examine: Use glob for ELF detection
Detection of ELF files fails where `setuid, setgid` is set on a file. By
allowing for extra output to be present in the `file` output, it will ensure
these files are still detected as binaries and dynamic libraries. This means
that setuid/getuid files are now stripped and have needed libraries added as
dependencies.