Check if sccache exists before enabling it
Now `ypkg` will check if `sccache` exists before attempting to
enabling it. Note that we don't check in the PATH environment
variable because that's not how `sccache` work in `ypkg`. Instead, we
check if `/usr/bin/sccache` exists. If it doesn't, `ypkg` will disable
`sccache` regardless of the build configuration and emit a warning.

Signed-off-by: Gavin Zhao <git@gzgz.dev>