Add PoC for compressing man pages
Compress man pages by default using a script from LFS.
compressdocs.sh does some hardlink and symlink checking as well as
checking if the man pages are already compressed rather than
just blindly gzip'ing all man pages which is why we're using it for now.

This works as-is, however, we may want to drop compressdocs.sh with
our own implementation in python. It will need to:

- Determine man page files or just walk $installdir/usr/share/man $installdir/usr/man
- Skip compression if already compressed (look for .gz, .bz2 append)
- Compress with the highest level of gzip compression and append with .gz
- Look for hardlinks or symlinks and handle them as needed

the files are not compressed first