yupdate: Add an option to not bump the release number
Often there is situation where we first try to bump and rebuild
a package to resolve an issue, then, update it if a simple rebuild
does not suffice.

In that case we end up with a double bump which the packager
will have to manually resolve.

Just add a no bump option to allievate this.