Consolidate fat LTO option
clang supports flto=auto as an equivalent to -flto/-flto=full.
Drop -flto=%YJOBS% as packages built on the buildserver would contaminate
with -flto=24 when they were used to build other packages, with python
projects mostly to blame. We'll rely on auto to get the right amount of
jobs from make/ninja.