ypkg2: Only enable fakeroot during the install & check steps
This will only allow $installdir to be used during install and check steps.
(Though, packagers should only be using it during install.)

This is extremely desirable due to the extreme overhead of fakeroot.
Especially when ccache is enabled but there are no cache hits for a build.

Unsure how yet if this will break builds in subtle ways. If needed, we can
add a context key to enable/disable this behavior.