ypkg-install-deps: add --dry-run and --json
The JSON dump has the following format:
```json
[
{
    "name": "<package-name>",
    "version": "<version>",
    "release": "<relno>",
    "packageHash": "<hash-of-eopkg>",
}
]
```

This not only dumps every single (including transitive) dependencies, but also
dumps the **currently installed** packages on the system. If we run this
inside the `solbuild` chroot and run `sudo eopkg up` prior to running `ypkg-
install-deps --dry-run --json package.yml`, we should get a (mostly)
reproducible JSON of packages that the root is comprised of. If we cache this
root identified by the hash of this JSON dump, we should be able to re-use
this cached root in `solbuild`, eliminating most if not all the time needed
for dependencies to install given that no dependency is changed.

Signed-off-by: Gavin Zhao <git@gzgz.dev>