ypkg2/scripts: Overhaul macro system
This (I hope) makes the ypkg macro handling story much easier to follow.
By adding dedicated classes for macros and actions, we have greater
flexibility in managing the system. It also gives the script generator
less to do by shifting the YAML parsing responsibility to the dedicated
Macros class.

As a part of this, I've also introduced the concept of arches, similar
to Boulder's arch/macro system. The goal of this is to have as much as
possible be defined in YAML files, getting them out of the Python code.
That will make it much easier to tweak our profiles as time goes on, and
hopefully makes it easier to switch to Boulder in the Eventual Times.

Also added are descriptions for actionable macros. While this isn't used
for anything yet, it enables us to add a CLI command to list and describe
them.

Note that the arches story isn't completely fleshed out yet. I just
wanted to start making commits before I ended up with one massive
changeset.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>