macros,scripts: Support tuning and flag options
Brand new territory for ypkg. This implements parsing for tuning groups
and flag groups in the macro files. The expected structure is that of
the AerynOS macro definition files to ease porting between the two
projects.

It is important to note that as of this commit, none of the flags are
actually used for the generated build scripts. That will be the next
task.

You might see that there is a lack of comments. That is because my brain
is fried, and I want to push this upstream so other people can start
looking at it. I'll add comments to All The Things another day. The YAML
parsing code might also make your eyes bleed. Sorry about that. But, at
least it parses everything without errors. I also haven't included the
arch file(s) yet because I'm just using the Aeryn base file for testing
the parsing.

The most amazing part is that I can actually build Nano with this
commit.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>